+++ housing_hunter_hero/graphics.py
@@ -1,4 +1,4 @@
-from housing_hunter import name_url_scraper, bed_bath_scraper, smash_together
+from housing_hunter_old import name_url_scraper, bed_bath_scraper, smash_together
 
 welcome_menu = """
