+++ housing_hunter_hero/housing_hunter.py
@@ -4,33 +4,33 @@ import json
 import re
 
 
-def name_url_scraper():
-
+def zip_scraper(zip_):
 	headers = {'User-Agent': 'Mozilla/5.0'}
-	url = "https://www.redfin.com/city/16163/WA/Seattle"
+	url = f'https://www.redfin.com/zipcode/{zip_}'
 	response = requests.get(url, headers=headers)
 
 	soup = BeautifulSoup(response.text, "html.parser")
 	soup = soup.find_all('script', type="application/ld+json")
-
+	bucket = []
 	x = []
 	for num in range(1, len(soup)):
 		x.append(json.loads(soup[num].get_text()))
 
-	lst = [v for i, v in enumerate(x) if i % 2 == 0]
-
-	while len(lst) > 10:
+	purge_ = [elm for elm in x if isinstance(elm, list)]
+	lst = [v for i, v in enumerate(purge_) if i % 2 == 0]
+	while len(lst) > 15:
 		lst.pop()
-	bucket = []
+
 	for item in lst:
-		bucket.append([item[0]['name'], item[1]['url']])
+		# print(type(item))
+		bucket.append([item[0]['name'].replace(zip_, ""), item[1]['url']])
 	return bucket
 
 
-def bed_bath_scraper():
-
+def bed_bath_scraper(zip_):
+	bucket = zip_scraper(zip_)
 	headers = {'User-Agent': 'Mozilla/5.0'}
-	url = "https://www.redfin.com/city/16163/WA/Seattle"
+	url = f"https://www.redfin.com/zipcode/{zip_}"
 	response = requests.get(url, headers=headers)
 
 	soup_ = BeautifulSoup(response.text, "html.parser")
@@ -50,9 +50,9 @@ def bed_bath_scraper():
 	for i in a_links:
 		return_me.append(i.parent.text)
 
-	while len(return_me) > 10:
+	while len(return_me) > 15:
 		return_me.pop()
-	while len(return_b) > 10:
+	while len(return_b) > 15:
 		return_b.pop()
 
 	for i in range(len(return_me)):
@@ -60,19 +60,33 @@ def bed_bath_scraper():
 		return_me[i] = return_me[i].replace("Bed s", "Beds ")
 		return_me[i] = return_me[i].replace("Bath", "Bath ")
 		return_me[i] = return_me[i].replace("Bath s", "Baths ")
-		return_me[i] = return_me[i].replace(return_b[i], " "+return_b[i]+" ")
+		return_me[i] = return_me[i].replace(return_b[i], " " + return_b[i] + " ")
 
 	for i in range(len(return_me)):
-		return_me[i] = re.sub(r'Listing(.*)', '', return_me[i])
+		return_me[i] = re.sub(f"{zip_}(.*)", '', return_me[i])
 
 	return return_me
 
 
-def smash_together(thing1, thing2):
-	for i in range(len(thing2)):
-		print(f"Property {i+1}: ", thing2[i], " - ", "\n ", f"https://www.redfin.com/{(thing1[i][1])}")
+def smash_together(list_1, list_2):
+	while list_2:
+		for x in range(len(list_1)):
+			if list_2[0].endswith(list_1[x][0]):
+				list_1[x].append(list_2[0])
+		list_2.pop(0)
+	results = []
+	for i in list_1:
+		if len(i) >= 3:
+			results.append(i)
+
+	return results
+
+
+def results_of_scrape(scraped_data):
+	for i in range(len(scraped_data)):
+		print(f"Property {i + 1}: ", scraped_data[i][2], "\n ", f"https://www.redfin.com/{scraped_data[i][1]}")
 		print(" ")
 
 
 if __name__ == "__main__":
-	smash_together(name_url_scraper(), bed_bath_scraper())
+	results_of_scrape((smash_together(zip_scraper('06118'), bed_bath_scraper('06118'))))
